requirements.txt
selenium
beautifulsoup4
lxml
google-generativeai
python-dotenv
colorama
webdriver-manager 